fix(api): satisfy lint after goal control additions
Move GoalControlWorkspaceAgentSession into its own file so
daemon_agent_sessions.go stays under the 800-line lint limit, and drop
the unused receiver name on fakeRuntime.GoalControl to satisfy both
revive and staticcheck.
